Lab_simulacion-main/proyecto/requirements.txt
pygame==2.5.2
numpy==1.26.4
matplotlib==3.10.1
PyYAML==6.0.2


